requirements.txt
requests==2.20.0
waiting==1.4.1
retry==0.9.2
tqdm==4.46.0
pyyaml==5.3.1
boto3==1.13.11
pytest=5.4.3
paramiko=2.7.1
libvirt-python=4.5.0
